Implement MarketplaceActivity fully
