+++ app/src/main/AndroidManifest.xml
@@ -104,7 +104,7 @@
         </activity>
 
         <activity
-            android:name=".MarketplaceActivity"
+            android:name=".gameActivities.MarketplaceActivity"
             android:parentActivityName=".TravelActivity"
             android:windowSoftInputMode="adjustPan">
