+++ app/src/main/res/layout/battleground_page2.xml
@@ -13,7 +13,7 @@
         android:id="@+id/agg_level"
         style="@style/user_stat_style3"
         android:layout_marginStart="8dp"
-        android:text="LEVEL 20"
+        android:text="LEVEL"
         android:textSize="25sp"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
@@ -26,7 +26,7 @@
         android:id="@+id/agg_level_progress"
         style="@style/user_stat_style3"
         android:layout_marginTop="48dp"
-        android:text="56 / 70"
+        android:text="LOADING..."
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.022"
@@ -37,7 +37,7 @@
     <TextView
         android:id="@+id/wood_gear_level"
         style="@style/user_stat_style3"
-        android:text="5"
+        android:text=""
         app:layout_constraintBottom_toTopOf="@+id/wood_res"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.441"
@@ -48,7 +48,7 @@
     <TextView
         android:id="@+id/fish_gear_level"
         style="@style/user_stat_style3"
-        android:text="100"
+        android:text=""
         app:layout_constraintBottom_toTopOf="@+id/fish_res"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.689"
@@ -59,7 +59,7 @@
     <TextView
         android:id="@+id/combat_gear_level"
         style="@style/user_stat_style3"
-        android:text="20"
+        android:text=""
         app:layout_constraintBottom_toTopOf="@+id/gold_res"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.936"
@@ -70,7 +70,7 @@
     <TextView
         android:id="@+id/wood_res"
         style="@style/user_stat_style4"
-        android:text="5"
+        android:text=""
         android:textColor="@color/final_green"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
@@ -82,7 +82,7 @@
     <TextView
         android:id="@+id/fish_res"
         style="@style/user_stat_style4"
-        android:text="100"
+        android:text=""
         android:textColor="@color/final_blue"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
@@ -94,7 +94,7 @@
     <TextView
         android:id="@+id/gold_res"
         style="@style/user_stat_style4"
-        android:text="20"
+        android:text=""
         android:textColor="@color/final_gold"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
