+++ app/src/main/res/layout/buttery_trading_hub.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MarketplaceActivity">
+    tools:context=".gameActivities.MarketplaceActivity">
 
     <Button
         android:id="@+id/tradeButton"
