+++ app/src/main/res/layout/marketplace_page2.xml
@@ -7,7 +7,7 @@
     android:id="@+id/marketplace_page2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MarketplaceActivity"
+    tools:context=".gameActivities.MarketplaceActivity"
     android:background="@drawable/design_2"
     android:orientation="vertical"
     >
@@ -16,7 +16,7 @@
         android:id="@+id/agg_level"
         style="@style/user_stat_style3"
         android:layout_marginStart="8dp"
-        android:text="LEVEL 20"
+        android:text="LEVEL"
         android:textSize="25sp"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
@@ -29,7 +29,7 @@
         android:id="@+id/agg_level_progress"
         style="@style/user_stat_style3"
         android:layout_marginTop="48dp"
-        android:text="56 / 70"
+        android:text="LOADING..."
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.022"
@@ -40,7 +40,7 @@
     <TextView
         android:id="@+id/wood_gear_level"
         style="@style/user_stat_style3"
-        android:text="5"
+        android:text=""
         app:layout_constraintBottom_toTopOf="@+id/wood_res"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.441"
@@ -51,7 +51,7 @@
     <TextView
         android:id="@+id/fish_gear_level"
         style="@style/user_stat_style3"
-        android:text="100"
+        android:text=""
         app:layout_constraintBottom_toTopOf="@+id/fish_res"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.689"
@@ -62,7 +62,7 @@
     <TextView
         android:id="@+id/combat_gear_level"
         style="@style/user_stat_style3"
-        android:text="20"
+        android:text=""
         app:layout_constraintBottom_toTopOf="@+id/gold_res"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.936"
@@ -73,7 +73,7 @@
     <TextView
         android:id="@+id/wood_res"
         style="@style/user_stat_style4"
-        android:text="5"
+        android:text=""
         android:textColor="@color/final_green"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
@@ -85,7 +85,7 @@
     <TextView
         android:id="@+id/fish_res"
         style="@style/user_stat_style4"
-        android:text="100"
+        android:text=""
         android:textColor="@color/final_blue"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
@@ -97,7 +97,7 @@
     <TextView
         android:id="@+id/gold_res"
         style="@style/user_stat_style4"
-        android:text="20"
+        android:text=""
         android:textColor="@color/final_gold"
         app:layout_constraintBottom_toTopOf="@+id/title"
         app:layout_constraintEnd_toEndOf="parent"
